created DependencyProvider and moved clients inside
