--- app/src/main/java/com/zooro/mvvmnewsapp/data/api/RetrofitInstance.kt
@@ -1,29 +1,0 @@
-package com.zooro.mvvmnewsapp.data.api
-
-import com.zooro.mvvmnewsapp.data.api.ApiSettings.Companion.BASE_URL
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
-import retrofit2.converter.gson.GsonConverterFactory
-
-class RetrofitInstance {
-    companion object{
-
-        private val retrofit by lazy {
-            val logging = HttpLoggingInterceptor()
-            logging.setLevel(HttpLoggingInterceptor.Level.BODY)
-            val client = OkHttpClient.Builder()
-                .addInterceptor(logging)
-                .build()
-            Retrofit.Builder()
-                .baseUrl(BASE_URL)
-                .addConverterFactory(GsonConverterFactory.create())
-                .client(client)
-                .build()
-        }
-
-        val api by lazy{
-            retrofit.create(NewsApiService::class.java)
-        }
-    }
-}
