--- app/src/main/java/com/zooro/mvvmnewsapp/di/Clients.kt
@@ -1,4 +1,0 @@
-package com.zooro.mvvmnewsapp.di
-
-object Clients {
-}
